+++ src/deptree/_meta.py
@@ -10,7 +10,10 @@ import importlib_metadata
 
 PROJECT_NAME = 'deptree'
 
-VERSION = importlib_metadata.version(PROJECT_NAME)
+_DISTRIBUTION_METADATA = importlib_metadata.metadata(PROJECT_NAME)
+
+SUMMARY = _DISTRIBUTION_METADATA['Summary']
+VERSION = _DISTRIBUTION_METADATA['Version']
 
 
 # EOF